Added another AI , and an ActionObject
The AI has a method  with the current middle as an Argument , will
return an ActionObject in result. Will either pass if current value on
hand is > 29 , change all 3 cards if it will increase the value. Also
determines via 'bruteforcing' which change is the most benefiting one.
Subject to change a lot.